Switched to MIT license
The code was previously released under CC BY 4.0. If you happened to use the code under CC BY 4.0, just continue doing so. If you update to the new version or use parts of the new code, just know, the code is partially (or fully, at your choice) licensed under MIT. Coming versions of the code are licensed under MIT, which should mean almost the same to you. CC is just unsuited for licensing code, which is why I switched.